broke up table.ts file in to its proper structure
