+++ src/administration/types/db-admin/create-keyspace.ts
@@ -34,7 +34,7 @@ import { AdminBlockingOptions } from '@/src/administration/types';
  *   updateDbKeyspace: true,
  * });
  *
- * // Creates collection in keyspace `new_keyspace` by default now
+ * // Creates collections in keyspace `new_keyspace` by default now
  * const coll = db.createCollection('my_coll');
  * ```
  *
