+++ src/db/types/collections/create-collection.ts
@@ -25,14 +25,14 @@ export interface CreateCollectionCommand {
 }
 
 /**
- * Options for creating a new collection.
+ * Options for creating a new collections.
  *
- * @field vector - The vector configuration for the collection.
- * @field indexing - The indexing configuration for the collection.
- * @field defaultId - The default ID for the collection.
- * @field keyspace - Overrides the keyspace for the collection.
+ * @field vector - The vector configuration for the collections.
+ * @field indexing - The indexing configuration for the collections.
+ * @field defaultId - The default ID for the collections.
+ * @field keyspace - Overrides the keyspace for the collections.
  * @field maxTimeMS - The maximum time to allow the operation to run.
- * @field checkExists - Whether to check if the collection exists before creating it.
+ * @field checkExists - Whether to check if the collections exists before creating it.
  *
  * @see Db.createCollection
  *
@@ -40,10 +40,10 @@ export interface CreateCollectionCommand {
  */
 export interface CreateCollectionOptions<Schema extends SomeDoc> extends WithTimeout, CollectionOptions<Schema>, CollectionSpawnOptions {
   /**
-   * If `true` or unset, runs an additional existence check before creating the collection, failing if the collection
+   * If `true` or unset, runs an additional existence check before creating the collections, failing if the collections
    * with the same name already exists, raising a {@link CollectionAlreadyExistsError}.
    *
-   * Otherwise, if `false`, the creation is always attempted, and the command will succeed even if the collection
+   * Otherwise, if `false`, the creation is always attempted, and the command will succeed even if the collections
    * with the given name already exists, as long as the options are the exact same (if options mismatch, it'll
    * throw a {@link DataAPIResponseError}).
    *
