+++ src/documents/index.ts
@@ -16,7 +16,9 @@ export * from './cursor';
 export * from './events';
 export * from './ids';
 export * from './collections';
+export * from './tables';
 export * from './embedding-providers';
+export * from './types';
 
 export {
   DataAPIResponseError,
