+++ src/documents/tables/types/row.ts
@@ -1,0 +1,32 @@
+// Copyright DataStax, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { SomeDoc } from "@/src/documents";
+
+export declare const $PrimaryKeyType: unique symbol;
+
+export interface Row<Schema extends SomeDoc, PrimaryKey extends (keyof Schema)[]> {
+  [$PrimaryKeyType]?: {
+    [P in PrimaryKey[number]]: Schema[P];
+  };
+}
+
+export type SomeTableKey = Record<string, unknown>;
+
+export type TableKey<Schema extends SomeDoc> = Schema extends { [$PrimaryKeyType]?: infer PrimaryKey }
+  ? PrimaryKey extends SomeTableKey
+    ? PrimaryKey
+    : SomeTableKey
+  : SomeTableKey;
+
